优化CreatorCenter工具函数
